finished crud operation the crime marks api and the crime types api

+++ backend/Application/DTOs/Responses/ShowOnMapCrimeResponse.cs
@@ -1,5 +1,5 @@
 ﻿namespace Application.DTOs.Responses
 {
-    public record ShowOnMapCrimeResponse(Guid Id, string CrimeTypeTitle, string? Location, double XPoint, double YPoint);
+    public record ShowOnMapCrimeResponse(Guid Id, string CrimeTypeTitle, string? Location, decimal XPoint, decimal YPoint);
 
 }
